# ml/openapi.json
{
  "components": {
    "schemas": {
      "ArmState": {
        "description": "One arm's posterior, as reported by /recommend.",
        "example": {
          "alpha": 7.4,
          "arm": "instagram_morning",
          "beta": 3.1,
          "channel": "instagram",
          "mean": 0.7048,
          "observedMean": 0.7,
          "timeSlot": "morning",
          "totalPulls": 10,
          "totalReward": 7.0
        },
        "properties": {
          "alpha": {
            "title": "Alpha",
            "type": "number"
          },
          "arm": {
            "title": "Arm",
            "type": "string"
          },
          "beta": {
            "title": "Beta",
            "type": "number"
          },
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "mean": {
            "title": "Mean",
            "type": "number"
          },
          "observedMean": {
            "title": "Observedmean",
            "type": "number"
          },
          "timeSlot": {
            "title": "Timeslot",
            "type": "string"
          },
          "totalPulls": {
            "title": "Totalpulls",
            "type": "integer"
          },
          "totalReward": {
            "title": "Totalreward",
            "type": "number"
          }
        },
        "required": [
          "arm",
          "channel",
          "timeSlot",
          "alpha",
          "beta",
          "mean",
          "observedMean",
          "totalPulls",
          "totalReward"
        ],
        "title": "ArmState",
        "type": "object"
      },
      "ClassifyRequest": {
        "example": {
          "sourceType": "Report",
          "text": "We deployed an upward-looking acoustic profiler through 340 m of first-year ice to measure krill swarms during the austral winter, sampling at 1 Hz for 72 hours."
        },
        "properties": {
          "sourceId": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceid"
          },
          "sourceType": {
            "default": "Report",
            "maxLength": 32,
            "minLength": 1,
            "title": "Sourcetype",
            "type": "string"
          },
          "text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ClassifyRequest",
        "type": "object"
      },
      "ClassifyResponse": {
        "example": {
          "confidenceScores": {
            "Biology / Ecology": 0.62,
            "Data Science / Instrumentation": 0.24
          },
          "modelVersion": "v1",
          "predictedTags": [
            "Biology / Ecology",
            "Data Science / Instrumentation"
          ]
        },
        "properties": {
          "confidenceScores": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Confidencescores",
            "type": "object"
          },
          "modelVersion": {
            "title": "Modelversion",
            "type": "string"
          },
          "predictedTags": {
            "items": {
              "type": "string"
            },
            "title": "Predictedtags",
            "type": "array"
          }
        },
        "required": [
          "predictedTags",
          "confidenceScores",
          "modelVersion"
        ],
        "title": "ClassifyResponse",
        "type": "object"
      },
      "ErrorBody": {
        "example": {
          "code": "INVALID_INPUT",
          "message": "text: String should have at least 1 character"
        },
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "description": "The three failure modes a caller has to be able to branch on.",
        "enum": [
          "INVALID_INPUT",
          "MODEL_NOT_LOADED",
          "INTERNAL"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecommendResponse": {
        "example": {
          "arm": "instagram_morning",
          "arms": [],
          "channel": "instagram",
          "expectedReward": 0.7048,
          "exploration": 0.8123,
          "reason": "highest Thompson sample over 9 arms",
          "timeSlot": "morning"
        },
        "properties": {
          "arm": {
            "title": "Arm",
            "type": "string"
          },
          "arms": {
            "items": {
              "$ref": "#/components/schemas/ArmState"
            },
            "title": "Arms",
            "type": "array"
          },
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "draftId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draftid"
          },
          "expectedReward": {
            "title": "Expectedreward",
            "type": "number"
          },
          "exploration": {
            "title": "Exploration",
            "type": "number"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "timeSlot": {
            "title": "Timeslot",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "timeSlot",
          "arm",
          "reason",
          "exploration",
          "expectedReward",
          "arms"
        ],
        "title": "RecommendResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Discipline auto-tagging for the IndoPolaris polar archive, plus the publish-time contextual bandit (Phase 7).",
    "title": "IndoPolaris ML Service",
    "version": "0.2.0",
    "x-generated-by": "ml/scripts/export_openapi.py"
  },
  "openapi": "3.1.0",
  "paths": {
    "/classify": {
      "post": {
        "description": "Auto-tag `text` (1\u20135000 characters) with IndoPolaris discipline tags.\n\nReturns `predictedTags` (at or above the confidence threshold) and\n`confidenceScores` for every label. An empty `predictedTags` is a valid\nanswer meaning \"no confident tag\" \u2014 the Node side then stores the record\nwith tagging pending rather than attaching a weak guess (Section 6).",
        "operationId": "classify_classify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed request body"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Model artifact not loaded"
          }
        },
        "summary": "Predict IndoPolaris discipline tags for a document"
      }
    },
...
    "/recommend": {
      "get": {
        "description": "Pick where and when to publish.\n\nSamples Beta(alpha, beta) once per arm and returns the argmax. Reads\n`BanditArmState` on every call \u2014 no in-memory cache \u2014 because the\nposteriors are the shared state between this service and any other\nwriter, and a cached copy would recommend from a snapshot the rest of\nthe system has already moved past.\n\n`exclude` is a comma-separated list of arms to hold back, so a simulator\nor an A/B harness can force exploration of a specific arm.\n\n`draftId` is echoed back and otherwise unused: it makes the response\ntraceable in a log, which is the only reason it exists.",
        "operationId": "recommend_recommend_get",
        "parameters": [
          {
            "in": "query",
            "name": "draftId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Draftid"
            }
          },
          {
            "in": "query",
            "name": "exclude",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exclude"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Thompson-sampled channel and time-slot for the next publish"
      }
    }
  }
}